requirements.txt
aiohttp==3.6.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
DateTime==4.3
discord==1.0.1
discord.py==1.5.1
Django==3.1.3
Flask==1.1.2
idna==2.10
image==1.5.33
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
Pillow==8.0.1
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
Werkzeug==1.0.1
yarl==1.5.1
zope.interface==5.2.0